# Dex*: Consistent field init order in constructors