--- a/sim/dex-abilities.ts
+++ b/sim/dex-abilities.ts
@@ -1,5 +1,5 @@
 import type {PokemonEventMethods, ConditionData} from './dex-conditions';
-import {BasicEffect, toID} from './dex-data';
+import {assignNewFields, BasicEffect, toID} from './dex-data';
 
 interface AbilityEventMethods {
 	onCheckShow?: (this: Battle, pokemon: Pokemon) => void;
@@ -38,7 +38,7 @@ export class Ability extends BasicEffect implements Readonly<BasicEffect> {
 	declare readonly condition?: ConditionData;
 
 	constructor(data: AnyObject) {
-		super(data);
+		super(data, false);
 
 		this.fullname = `ability: ${this.name}`;
 		this.effectType = 'Ability';
@@ -63,6 +63,7 @@ export class Ability extends BasicEffect implements Readonly<BasicEffect> {
 				this.gen = 3;
 			}
 		}
+		assignNewFields(this, data);
 	}
 }