Spinoff of https://github.com/smogon/pokemon-showdown/pull/10549

---

Rather simple change to get consistent field init order where possible.

I realized at the end that there's a very simple set of rules for making sure the ctors have the same behavior as they did with `Object.assign()`:
1) Don't reference `this.field` before it's been initialized
    it doesn't exist.
2) Don't reference `data.field` after `this.field` has been initialized
    you'll miss out on the default values and other fix-up logic.
3) Copy extra fields at the end

And... that's it.

---
# Results
## Memory
Heap snapshots, analyzed w/ Chrome's devtools.

Going by 'Constructor' and 'Alloc. Size' columns.

#### All mods
|     Type | Before |  After |
| -------- | -----: | -----: |
| Species  | 27059K | 25986K |
| (system) | **26360K** | **13444K** |
| DataMove | 17737K | 16312K |
|     Item |  6220K |  6085K |
|  Ability |  2893K |  2606K |
|   Format |   242K |   129K |
|   Nature |   155K |   155K |
| TypeInfo |    68K |    68K |

According to Chrome's UI, **total memory went from 279 to 260 MB**.

The point of consistent field ordering was to greatly reduce the number of hidden classes.
The 13M decrease in the `(system)` category indicates this worked.
Ta-da.

#### Just Dex
|     Type | Before | After |
| -------- | -----: | ----: |
| (system) |  3490K | 1836K |
| DataMove |  1473K |  731K |
| Species  |  1460K |  705K |
|     Item |   382K |  240K |
|  Ability |   326K |  188K |
|   Format |   242K |  129K |
|   Nature |   4200 |  4200 |
| TypeInfo |   2736 |  1824 |

According to Chrome's UI, total memory went from 53.6 to 50.0 MB.

## Times

My test runs were taken on a local branch with #10616  merged in.

|              Action | Before |   After |  Unit |
| ------------------- | ------ | ------- | ----- |
|          `npm test` |  11961 |   11591 | ticks |
| `npm run full-test` | 140903 |  132434 | ticks |
|        Dex.loadData |    246 |     248 |    ms |
|         Dex.*.all() |    120 |      67 |    ms |

Where '*' is [moves, items, species, abilities, types, natures]
Dex also loads Formats as part of loadData.

About 8 seconds saved on `npm run full-test`.

## Benchmark code

#### ticks for `npm` commands (`./package.json`)
```sh
mocha --v8-prof ...
# Note: enabling prof adds a few % overhead
```

#### milliseconds for Dex loading
```js
const start0 = performance.now();
Dex.loadData();
const start1 = performance.now();
Dex.moves.all();
Dex.items.all();
Dex.species.all();
Dex.abilities.all();
Dex.types.all();
Dex.natures.all();
const end = performance.now();
console.log('loadData', start1 - start0);
console.log('.alls', end - start1);
```

#### Heap snapshots
```js
const v8 = require('v8');
v8.writeHeapSnapshot(); // before
Dex.loadData();
Dex.moves.all();
Dex.items.all();
Dex.species.all();
Dex.abilities.all();
Dex.types.all();
Dex.natures.all();
v8.writeHeapSnapshot(); // after
```

---

Covers all relevant classes.
```
$ grep -Ern "extends BasicEffect" ./sim/ ./server/ ./tools/ ./test/ ./config/ ./data ./lib/
./sim/dex-items.ts:23:export class Item extends BasicEffect implements Readonly<BasicEffect> {
./sim/dex-abilities.ts:31:export class Ability extends BasicEffect implements Readonly<BasicEffect> {
./sim/dex-moves.ts:359:export class DataMove extends BasicEffect implements Readonly<BasicEffect & MoveData> {
./sim/dex-conditions.ts:613:export class Condition extends BasicEffect implements
./sim/dex-species.ts:87:export class Species extends BasicEffect implements Readonly<BasicEffect & SpeciesFormatsData> {
./sim/dex-formats.ts:388:export class Format extends BasicEffect implements Readonly<BasicEffect> {
./sim/dex-data.ts:150:export class Nature extends BasicEffect implements Readonly<BasicEffect & NatureData> {
```

---

<details><summary>old todo</summary>

## TODO:
Simple testing indicates that Formats has a lot of extra fields.
Makes sense when you check the class definition - it has many optional properties.
In that scenario, one could try sorting the keys before iterating over them.
You'd want to measure whether sorting the keys meaningfully reduces the number of unique sequences.
</details>
I've tested sorting the extra keys and it did not improve any of the above metrics. So, I won't do that.