@larry-the-table-guy 2024-10-23
What's weird is that some of those keys don't appear in the corresponding ./data files for the type. Like `onDeductPP` happening during Species, yet it only appears in `abilities` data files.

I don't know why that happens.