@larry-the-table-guy 2024-10-28
Came up with a much better way to analyze the # of object shapes.

---
## Old state (master)
`3324` total shapes, `3030` of which could be merged (by re-ordering keys) into a set of `429`. So, `2601` / `3324` shapes were avoidable.

```
# actual shapes per effectType Map(7) {
  'Format' => 80,
  'Pokemon' => 1265,
  'Move' => 1148,
  'Item' => 272,
  'Ability' => 551,
  'Type' => 6,
  'Nature' => 2
}
min possible # shapes per effectType Map(7) {
  'Format' => 24,
  'Pokemon' => 21,
  'Move' => 326,
  'Item' => 93,
  'Ability' => 257,
  'Type' => 1,
  'Nature' => 1
}
```

---
## current PR state
Among the dex* objects, there are `937` object shapes, `379` of which could be reduced into a set of `165`. So, `214` / `937` shapes are avoidable.

Breakdown by type:
```
# total object shapes: 937
# actual shapes per effectType Map(7) {
  'Format' => 38,
  'Pokemon' => 23,
  'Move' => 418,
  'Item' => 153,
  'Ability' => 303,
  'Type' => 1,
  'Nature' => 1
}
min possible # shapes per effectType Map(7) {
  'Format' => 24,
  'Pokemon' => 21,
  'Move' => 326,
  'Item' => 93,
  'Ability' => 257,
  'Type' => 1,
  'Nature' => 1
}
```

## With sorting
`741` shapes,  `33` of which could be reduced into `15`.

Those remaining extra shapes seem to come from the position of `zMove` and `maxMove` in `Move`, and `onTypePriority` & `onType` in `Pokemon`. Since sorting didn't remove those, there must be code paths in the constructors that don't init those fields in the same order as each other.
```
# actual shapes per effectType Map(7) {
  'Format' => 24,
  'Pokemon' => 22,
  'Move' => 343,
  'Item' => 93,
  'Ability' => 257,
  'Type' => 1,
  'Nature' => 1
}
min possible # shapes per effectType Map(7) {
  'Format' => 24,
  'Pokemon' => 21,
  'Move' => 326,
  'Item' => 93,
  'Ability' => 257,
  'Type' => 1,
  'Nature' => 1
}
```

---

Conclusion: will be put in a later comment

---

<details><summary>setup code</summary>

(bottom of `dex.ts`)
```ts
Dex.includeMods();
let obj_list = [...Dex.formats.all()];
for (const d in dexes) {
        if (d === 'base') continue;
        const m = dexes[d];
        // slow but w/e
        obj_list = obj_list.concat(
                m.species.all(),
                m.moves.all(),
                m.items.all(),
                m.abilities.all(),
                m.types.all(),
                m.natures.all()
        );
}
const key_seqs = new Map();
for (const obj of obj_list) {
        const sorted = obj.effectType + '|' + Object.keys(obj).sort().join(';');
        const keys = Object.keys(obj).join(';');
        let bucket = key_seqs.get(sorted);
        if (!bucket) key_seqs.set(sorted, bucket = new Map());
        bucket.set(keys, (bucket.get(keys) || 0) + 1);
}

let num_total_shapes = 0;
for (const v of key_seqs.values()) num_total_shapes += v.size;
console.log("# total object shapes:", num_total_shapes);

const num_act_shapes_per_effectType = new Map();
const num_min_shapes_per_effectType = new Map();
for (const [k, v] of key_seqs.entries()) {
        const key = k.split('|')[0];
        num_act_shapes_per_effectType.set(key, (num_act_shapes_per_effectType.get(key) || 0) + v.size);
        num_min_shapes_per_effectType.set(key, (num_min_shapes_per_effectType.get(key) || 0) + 1);
}
console.log("# actual shapes per effectType", num_act_shapes_per_effectType);
console.log("min possible # shapes per effectType", num_min_shapes_per_effectType);

for (const k of key_seqs.keys()) {
        if (key_seqs.get(k).size === 1) key_seqs.delete(k);
}
console.log(key_seqs);

console.log("# REDUNDANT SHAPES PER KEYSET");
let num_redundant_shapes = 0;
for (const [k, v] of key_seqs.entries()) {
        console.log(k, v.size);
        num_redundant_shapes += v.size;
}
console.log("# TOTAL REDUNDANT SHAPES, uniq vs actual:", key_seqs.size, ",", num_redundant_shapes);

```
</details>