@Slayer95 2024-10-29
> What's weird is that some of those keys don't appear in the corresponding ./data files for the type. Like `onDeductPP` happening during Species, yet it only appears in `abilities` data files.
> 
> I don't know why that happens.

This is likely the doing of a deleted mod which hasn't been correspondingly deleted from the built files. Recreate the dist folder, probably?