@larry-the-table-guy 2024-10-30
> > What's weird is that some of those keys don't appear in the corresponding ./data files for the type. Like `onDeductPP` happening during Species, yet it only appears in `abilities` data files.
> > I don't know why that happens.
> 
> This is likely the doing of a deleted mod which hasn't been correspondingly deleted from the built files. Recreate the dist folder, probably?

So, the issue goes away when I group the extra fields by `effectType` (as opposed to relying on chronology), which led me to think that the loading order is not what I thought it was. And indeed it wasn't.
https://github.com/smogon/pokemon-showdown/blob/5a232233fec290cffb2f5f96903f0ddbf0850165/sim/dex-species.ts#L566

Well, that's good to know for the future.